Fix enabled check tests
